+++ src/mako/security/crypto/CryptoManager.php
@@ -8,6 +8,7 @@
 namespace mako\security\crypto;
 
 use mako\common\AdapterManager;
+use mako\security\crypto\encrypters\EncrypterInterface;
 use mako\security\crypto\encrypters\OpenSSL;
 use mako\security\Key;
 use mako\security\Signer;
@@ -20,6 +21,7 @@ use function vsprintf;
  *
  * @mixin \mako\security\crypto\encrypters\EncrypterInterface
  * @method \mako\security\crypto\encrypters\EncrypterInterface instance(?string $configuration = null)
+ * @method \mako\security\crypto\encrypters\EncrypterInterface getInstance(?string $configuration = null)
  */
 class CryptoManager extends AdapterManager
 {
@@ -50,5 +52,16 @@ class CryptoManager extends AdapterManager
 		$configuration = $this->configurations[$configuration];
 
 		return new Crypto($this->factory($configuration['library'], $configuration), $this->container->get(Signer::class));
+	}
+
+	/**
+	 * Returns an instance of the chosen encrypter. Alias of CryptoManager::getInstance().
+	 *
+	 * @param  string|null                                        $configuration Configuration name
+	 * @return mako\security\crypto\encrypters\EncrypterInterface
+	 */
+	public function getEncrypter(?string $configuration = null): EncrypterInterface
+	{
+		return $this->getInstance($configuration);
 	}
 }